requirements/base.txt
django
djangorestframework
djangorestframework-simplejwt
drf-yasg
requests
python-dotenv
websocket-client
psycopg2
drf-spectacular
django-cors-headers
setuptools

# analyze
pandas
pandas_ta
django-pandas
numpy==1.26.4

# tasks
redis
celery
django-celery-beat

# apis
python-binance